frontend: Moved UpdatePopup into file with own name
