+++ frontend/src/App.tsx
@@ -35,8 +35,8 @@ import { ClusterTitle } from './components/cluster/Chooser';
 import ActionsNotifier from './components/common/ActionsNotifier';
 import AlertNotification from './components/common/AlertNotification';
 import ReleaseNotesModal from './components/common/ReleaseNotes/ReleaseNotesModal';
+import UpdatePopup from './components/common/ReleaseNotes/UpdatePopup';
 import Sidebar, { drawerWidth, NavigationTabs, useSidebarItem } from './components/Sidebar';
-import UpdatePopup from './components/updatepopup';
 import helpers from './helpers';
 import { useElectronI18n } from './i18n/electronI18n';
 import LocaleSelect from './i18n/LocaleSelect/LocaleSelect';
